Extract combatant UI into components
Split src/routes/+page.svelte (1013 lines → 279) into five
presentational components in a new src/components/ layer:

- TopBar.svelte: encounter name, phase, round, active turn
- SetupPanel.svelte: creature picker, template control, preview,
  custom combatant form, start/reset
- InitiativeTrack.svelte: ordered initiative list + unordered queue
- CombatantCard.svelte: HP row, HP track, damage/heal/temp/set-zero
- FeedbackPanel.svelte: severity-styled event feedback list

Components own their own form scratch state and emit intentions via
callback props. The route keeps encounter state, command dispatch,
and shared HP/temp inputs.

Update CLAUDE.md architecture diagram to add src/components/ as a
layer between src/routes/ and src/lib/.

No behavior change. Pre-PR gate (check, test:run, audit, build) all
pass; manual smoke test recommended in dev server.

Part of #15 (slice 1 of 7).
Closes #28.
